Fix major compilation issues in ILGPU.Algorithms
- Add missing IntrinsicMath methods (Sqrt, Atan2, PI)
- Add ToArray() extension method for ArrayView<T>
- Add ArrayViewDenseExtensions for Dense stride operations
- Add [Kernel] attributes to all kernel methods
- Convert FFTPlan from class to struct
- Fix tuple Dispose() error by disposing components
- Fix GetValueOrDefault with TryGetValue pattern
- Fix Finalize() method warning in MPINative
- Fix ambiguous conditional expressions
- Improve FFT kernel signatures and implementations

+++ Src/ILGPU.Algorithms/Distributed/DistributedAlgorithms.cs
@@ -495,7 +495,7 @@ namespace ILGPU.Algorithms.Distributed
             var totalSize = 0;
             foreach (var partition in partitions)
             {
-                totalSize += partition.Length;
+                totalSize += (int)partition.Length;
             }
             
             return mpiAccelerator.LocalAccelerator.Allocate1D<T>(totalSize).View;
